pyserver/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.2
openai>=1.40
psycopg[binary]>=3.1
passlib[bcrypt,argon2]>=1.7
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
email-validator>=2.1
httpx>=0.27
